requirements.txt
openai
gradio
textblob
python-dotenv